+++ example/logic_test.sh
@@ -4,23 +4,23 @@ ROOT_DIR="$(dirname "${BASH_SOURCE[0]}")/.."
 SCRIPT="$ROOT_DIR/example/logic.sh"
 
 function test_text_should_be_equal() {
-  assertEquals "expected 123" "$($SCRIPT "123")"
+  assert_equals "expected 123" "$($SCRIPT "123")"
 }
 
 function test_text_should_contain() {
-  assertContains "expect" "$($SCRIPT "123")"
+  assert_contains "expect" "$($SCRIPT "123")"
 }
 
 function test_text_should_not_contain() {
-  assertNotContains "expecs" "$($SCRIPT "123")"
+  assert_not_contains "expecs" "$($SCRIPT "123")"
 }
 
 function test_text_should_match_a_regular_expression() {
-  assertMatches ".*xpec*" "$($SCRIPT "123")"
+  assert_matches ".*xpec*" "$($SCRIPT "123")"
 }
 
 function test_text_should_not_match_a_regular_expression() {
-  assertNotMatches ".*xpes.*" "$($SCRIPT "123")"
+  assert_not_matches ".*xpes.*" "$($SCRIPT "123")"
 }
 
 function test_should_validate_an_ok_exit_code() {
@@ -30,7 +30,7 @@ function test_should_validate_an_ok_exit_code() {
 
   fake_function
 
-  assertExitCode "0"
+  assert_exit_code "0"
 }
 
 
@@ -41,7 +41,7 @@ function test_should_validate_a_non_ok_exit_code() {
 
   fake_function
 
-  assertExitCode "1"
+  assert_exit_code "1"
 }
 
 function test_other_way_of_using_the_exit_code() {
@@ -49,7 +49,7 @@ function test_other_way_of_using_the_exit_code() {
     return 1
   }
 
-  assertExitCode "1" "$(fake_function)"
+  assert_exit_code "1" "$(fake_function)"
 }
 
 function test_successful_exit_code() {
@@ -57,7 +57,7 @@ function test_successful_exit_code() {
     return 0
   }
 
-  assertSuccessfulCode "$(fake_function)"
+  assert_successful_code "$(fake_function)"
 }
 
 function test_other_way_of_using_the_successful_exit_code() {
@@ -67,7 +67,7 @@ function test_other_way_of_using_the_successful_exit_code() {
 
   fake_function
 
-  assertSuccessfulCode
+  assert_successful_code
 }
 
 function test_general_error() {
@@ -75,7 +75,7 @@ function test_general_error() {
     return 1
   }
 
-  assertGeneralError "$(fake_function)"
+  assert_general_error "$(fake_function)"
 }
 
 function test_other_way_of_using_the_general_error() {
@@ -85,21 +85,21 @@ function test_other_way_of_using_the_general_error() {
 
   fake_function
 
-  assertGeneralError
+  assert_general_error
 }
 
 function test_should_assert_exit_code_of_a_non_existing_command() {
-  assertCommandNotFound "$(a_non_existing_function > /dev/null 2>&1)"
+  assert_command_not_found "$(a_non_existing_function > /dev/null 2>&1)"
 }
 
 function test_should_assert_that_an_array_contains_1234() {
   local distros=(Ubuntu 1234 Linux\ Mint)
 
-  assertArrayContains "1234" "${distros[@]}"
+  assert_array_contains "1234" "${distros[@]}"
 }
 
 function test_should_assert_that_an_array_not_contains_1234() {
   local distros=(Ubuntu 1234 Linux\ Mint)
 
-  assertArrayNotContains "a_non_existing_element" "${distros[@]}"
+  assert_array_not_contains "a_non_existing_element" "${distros[@]}"
 }