+++ tests/functional/logic_test.sh
@@ -5,23 +5,23 @@ ROOT_DIR="$(dirname "${BASH_SOURCE[0]}")"
 SCRIPT="$ROOT_DIR/logic.sh"
 
 function test_text_should_be_equal() {
-  assertEquals "expected 123" "$($SCRIPT "123")"
+  assert_equals "expected 123" "$($SCRIPT "123")"
 }
 
 function test_text_should_contain() {
-  assertContains "expect" "$($SCRIPT "123")"
+  assert_contains "expect" "$($SCRIPT "123")"
 }
 
 function test_text_should_not_contain() {
-  assertNotContains "expecs" "$($SCRIPT "123")"
+  assert_not_contains "expecs" "$($SCRIPT "123")"
 }
 
 function test_text_should_match_a_regular_expression() {
-  assertNotContains ".*xpec*" "$($SCRIPT "123")"
+  assert_not_contains ".*xpec*" "$($SCRIPT "123")"
 }
 
 function test_text_should_not_match_a_regular_expression() {
-  assertNotContains ".*xpes*" "$($SCRIPT "123")"
+  assert_not_contains ".*xpes*" "$($SCRIPT "123")"
 }
 
 function test_should_validate_an_ok_exit_code() {
@@ -31,7 +31,7 @@ function test_should_validate_an_ok_exit_code() {
 
   fake_function
 
-  assertExitCode "0"
+  assert_exit_code "0"
 }
 
 function test_should_validate_a_non_ok_exit_code() {
@@ -41,7 +41,7 @@ function test_should_validate_a_non_ok_exit_code() {
 
   fake_function
 
-  assertExitCode "1"
+  assert_exit_code "1"
 }
 
 function test_other_way_of_using_the_exit_code() {
@@ -49,5 +49,5 @@ function test_other_way_of_using_the_exit_code() {
     return 1
   }
 
-  assertExitCode "1" "$(fake_function)"
+  assert_exit_code "1" "$(fake_function)"
 }