+++ tests/unit/helpers_test.sh
@@ -13,25 +13,25 @@ function dummyFunction() {
 }
 
 function test_normalizeTestFunctionName_empty() {
-  assertEquals "" "$(Helper::normalizeTestFunctionName)"
+  assert_equals "" "$(Helper::normalizeTestFunctionName)"
 }
 
 function test_normalizeTestFunctionName_one_word() {
-  assertEquals "Word" "$(Helper::normalizeTestFunctionName "word")"
+  assert_equals "Word" "$(Helper::normalizeTestFunctionName "word")"
 }
 
 function test_normalizeTestFunctionName_snake_case() {
-  assertEquals "Some logic" "$(Helper::normalizeTestFunctionName "test_some_logic")"
+  assert_equals "Some logic" "$(Helper::normalizeTestFunctionName "test_some_logic")"
 }
 
 function test_normalizeTestFunctionName_camel_case() {
-  assertEquals "SomeLogic" "$(Helper::normalizeTestFunctionName "testSomeLogic")"
+  assert_equals "SomeLogic" "$(Helper::normalizeTestFunctionName "testSomeLogic")"
 }
 
 function test_getFunctionsToRun_no_filter_should_return_all_functions() {
   local functions=("prefix_function1" "prefix_function2" "other_function" "prefix_function3")
 
-  assertEquals\
+  assert_equals\
     "prefix_function1 prefix_function2 prefix_function3"\
     "$(Helper::getFunctionsToRun "prefix" "" "${functions[*]}")"
 }
@@ -39,35 +39,35 @@ function test_getFunctionsToRun_no_filter_should_return_all_functions() {
 function test_getFunctionsToRun_with_filter_should_return_matching_functions() {
   local functions=("prefix_function1" "prefix_function2" "other_function" "prefix_function3")
 
-  assertEquals "prefix_function1" "$(Helper::getFunctionsToRun "prefix" "function1" "${functions[*]}")"
+  assert_equals "prefix_function1" "$(Helper::getFunctionsToRun "prefix" "function1" "${functions[*]}")"
 }
 
 function test_getFunctionsToRun_filter_no_matching_functions_should_return_empty() {
   local functions=("prefix_function1" "prefix_function2" "other_function" "prefix_function3")
 
-  assertEquals "" "$(Helper::getFunctionsToRun "prefix" "nonexistent" "${functions[*]}")"
+  assert_equals "" "$(Helper::getFunctionsToRun "prefix" "nonexistent" "${functions[*]}")"
 }
 
 function test_getFunctionsToRun_fail_when_duplicates() {
   local functions=("prefix_function1" "prefix_function1")
 
-  assertGeneralError "$(Helper::getFunctionsToRun "prefix" "" "${functions[*]}")"
+  assert_general_error "$(Helper::getFunctionsToRun "prefix" "" "${functions[*]}")"
 }
 
 function test_dummyFunction_is_executed_with_execute_function_if_exists() {
   local function_name='dummyFunction'
 
-  assertEquals "dummyFunction executed" "$(Helper::executeFunctionIfExists "$function_name")"
+  assert_equals "dummyFunction executed" "$(Helper::executeFunctionIfExists "$function_name")"
 }
 
 function test_no_function_is_executed_with_execute_function_if_exists() {
   local function_name='notExistingFunction'
 
-  assertEmpty "$(Helper::executeFunctionIfExists "$function_name")"
+  assert_empty "$(Helper::executeFunctionIfExists "$function_name")"
 }
 
 function test_unsuccessful_unsetIfExists() {
-  assertGeneralError "$(Helper::unsetIfExists "fake_function")"
+  assert_general_error "$(Helper::unsetIfExists "fake_function")"
 }
 
 function test_successful_unsetIfExists() {
@@ -76,19 +76,19 @@ function test_successful_unsetIfExists() {
     return 0
   }
 
-  assertSuccessfulCode "$(Helper::unsetIfExists "fake_function")"
+  assert_successful_code "$(Helper::unsetIfExists "fake_function")"
 }
 
 function test_checkDuplicateFunctions_with_duplicates() {
   local file
   file="$(dirname "${BASH_SOURCE[0]}")/fixtures/duplicate_functions.sh"
 
-  assertGeneralError "$(Helper::checkDuplicateFunctions "$file")"
+  assert_general_error "$(Helper::checkDuplicateFunctions "$file")"
 }
 
 function test_checkDuplicateFunctions_without_duplicates() {
   local file
   file="$(dirname "${BASH_SOURCE[0]}")/fixtures/no_duplicate_functions.sh"
 
-  assertSuccessfulCode "$(Helper::checkDuplicateFunctions "$file")"
+  assert_successful_code "$(Helper::checkDuplicateFunctions "$file")"
 }