+++ tests/unit/setup_teardown_test.sh
@@ -19,9 +19,9 @@ function tearDownAfterScript() {
 }
 
 function test_counter_is_incremented_after_setup_before_script_and_setup() {
-  assertEquals "3" "$TEST_COUNTER"
+  assert_equals "3" "$TEST_COUNTER"
 }
 
 function test_counter_is_decremented_and_incremented_after_teardown_and_setup() {
-  assertEquals "3" "$TEST_COUNTER"
+  assert_equals "3" "$TEST_COUNTER"
 }