+++ CHANGELOG.md
@@ -3,9 +3,24 @@
 ## [Unreleased](https://github.com/TypedDevs/bashunit/compare/0.6.0...main)
 - `mock`
 - `spy`
-- `assertHaveBeenCalledWith`
-- `assertHaveBeenCalled`
-- `assertHaveBeenCalledTimes`
+- `assert_have_been_called`
+- `assert_have_been_called_with`
+- `assert_have_been_called_times`
+- Rename assertions from camelCase to snake_case:
+    - `assertEquals` -> `assert_equals`
+    - `assertNotEquals` -> `assert_not_equals`
+    - `assertEmpty` -> `assert_empty`
+    - `assertNotEmpty` -> `assert_not_empty`
+    - `assertContains` -> `assert_contains`
+    - `assertNotContains` -> `assert_not_contains`
+    - `assertMatches` -> `assert_matches`
+    - `assertNotMatches` -> `assert_not_matches`
+    - `assertExitCode` -> `assert_exit_code`
+    - `assertSuccessfulCode` -> `assert_successful_code`
+    - `assertGeneralError` -> `assert_general_error`
+    - `assertCommandNotFound` -> `assert_command_not_found`
+    - `assertArrayContains` -> `assert_array_contains`
+    - `assertArrayNotContains` -> `assert_array_not_contains`
 
 ## [0.6.0](https://github.com/TypedDevs/bashunit/compare/0.5.0...0.6.0) - 2023-09-19